Phosphor SVG Icons

Create a <svg></svg> tag to render an icon.

Usage

ph(
  name,
  weight = c("light", "regular", "thin", "bold", "fill"),
  fill = "currentColor",
  rotate = NULL,
  height = "1.33em",
  width = NULL,
  vertical_align = "-0.25em",
  title = name,
  ...
)

Arguments

name

Name of the icon to use.

weight

Weight of icon (from thinnest to thickest): thin, light (default), regular, bold or fill.

fill

Fill color for the icon, default is currentColor which should match the color of the text where the icon is used.

rotate

Numeric, angle to rotate the icon.

height, width

Height and width in valid CSS unit.

vertical_align

Vertical alignment for the icon, this depend on the size of the icon.

title

Add a <title> tag to provides an accessible, short-text description of the icon. Use NULL for no title.

...

Attributes passed to the SVG tag.

Value

An SVG tag.

Examples

library(phosphoricons)

ph("airplane-tilt")
#> <svg xmlns="http://www.w3.org/2000/svg" viewbox="0 0 256 256" class="phosphor-svg" height="1.33em" fill="currentColor" style="vertical-align:-0.25em;">
#>   <path d="M214.66089,83.53027a30.00065,30.00065,0,0,0-43.98633-40.74414,6.044,6.044,0,0,0-.811,1.00684l-27.6289,29.22168L57.93823,42.36133a5.998,5.998,0,0,0-6.29346,1.39648l-24,24a5.99966,5.99966,0,0,0,.91455,9.23438l65.91016,43.94043L77.4021,138H55.88745a5.99782,5.99782,0,0,0-4.24268,1.75781l-24,24a5.99956,5.99956,0,0,0,2.31983,9.92578l39,13.19532,13.24072,39.04785a5.99955,5.99955,0,0,0,9.9248,2.31543l24-24A5.99653,5.99653,0,0,0,117.88745,200V178.65527l17.126-15.55957,43.91993,64.28907a6.00027,6.00027,0,0,0,9.19677.85742l24-24a5.99849,5.99849,0,0,0,1.396-6.293L182.87768,113.667l30.76856-28.89062A6.00123,6.00123,0,0,0,214.66089,83.53027ZM204.615,76.72852q-.21534.21532-.40528.44726L171.78,107.626a6,6,0,0,0-1.53125,6.4248l30.69433,84.4082L184.7605,214.6416l-43.72657-64.00586a6,6,0,0,0-8.98926-1.05566l-24.19189,21.97949A5.99857,5.99857,0,0,0,105.88745,176v21.51465L90.47,212.93262,79.37719,180.21875a5.99977,5.99977,0,0,0-3.75927-3.75684L42.9602,165.41211,58.3728,150H79.88745a5.99778,5.99778,0,0,0,4.24267-1.75781l24-24a5.99965,5.99965,0,0,0-.91455-9.23438L41.30542,71.06738,57.428,54.94434l84.40918,30.69433a6.0016,6.0016,0,0,0,6.41015-1.5166L179.95776,50.583c.08545-.08985.16748-.18262.24609-.27637A18.00028,18.00028,0,0,1,204.615,76.72852Z"></path>
#>   <title>airplane-tilt-light</title>
#> </svg>
ph("airplane-tilt", weight = "thin")
#> <svg xmlns="http://www.w3.org/2000/svg" viewbox="0 0 256 256" class="phosphor-svg" height="1.33em" fill="currentColor" style="vertical-align:-0.25em;">
#>   <path d="M211.68628,44.20117a28.02993,28.02993,0,0,0-39.59766,0,3.99779,3.99779,0,0,0-.62891.81445L142.78589,75.34277,57.25464,44.24121a3.99914,3.99914,0,0,0-4.19581.93066l-24,24a3.99981,3.99981,0,0,0,.60987,6.15625l67.94043,45.29395L78.23071,140H55.88745a4.00164,4.00164,0,0,0-2.82862,1.17187l-24,24a4.00013,4.00013,0,0,0,1.54639,6.61719l39.936,13.51172,13.5581,39.9834a3.99974,3.99974,0,0,0,6.6167,1.54394l24-24A4,4,0,0,0,115.88745,200V177.77051l19.48144-17.69922,45.21582,66.18555a4.00062,4.00062,0,0,0,6.13135.57128l24-24a4.00075,4.00075,0,0,0,.93067-4.19531L180.5476,113.11133l31.7295-29.793a3.98575,3.98575,0,0,0,.75048-.957A28.03809,28.03809,0,0,0,211.68628,44.20117Zm-5.65674,33.94141a3.96731,3.96731,0,0,0-.34375.39062L173.14917,109.084a4.00086,4.00086,0,0,0-1.021,4.28321l31.1294,85.60547-18.78809,18.78808-45.08691-65.99707a3.99992,3.99992,0,0,0-2.83594-1.71582,3.9595,3.9595,0,0,0-.46729-.02734,4.00181,4.00181,0,0,0-2.68945,1.03906l-24.19238,21.97949A4.00232,4.00232,0,0,0,107.88745,176v22.34277L89.60913,216.62109l-12.126-35.75976a4.00019,4.00019,0,0,0-2.50586-2.50488l-35.708-12.082L57.54419,148H79.88745a4.00165,4.00165,0,0,0,2.82861-1.17188l24-24a3.9998,3.9998,0,0,0-.60986-6.15625L38.16577,71.37793l18.749-18.74805,85.60547,31.12891a3.99517,3.99517,0,0,0,4.27344-1.01074L178.50464,49.209c.0913-.09668.17724-.19726.25732-.30078a20.00067,20.00067,0,0,1,27.26758,29.23438Z"></path>
#>   <title>airplane-tilt-thin</title>
#> </svg>
ph("airplane-tilt", weight = "bold")
#> <svg xmlns="http://www.w3.org/2000/svg" viewbox="0 0 256 256" class="phosphor-svg" height="1.33em" fill="currentColor" style="vertical-align:-0.25em;">
#>   <path d="M189.86841,115.333l27.88476-26.18262a12.004,12.004,0,0,0,1.73779-2.03906A36.00061,36.00061,0,0,0,166.43188,38.544a12.06,12.06,0,0,0-1.25146,1.46875L140.58178,66.0293,59.988,36.72266a11.99848,11.99848,0,0,0-12.58593,2.792l-24,24a12.00005,12.00005,0,0,0,1.8291,18.46972l59.82031,39.88086L74.91675,132H55.88745a12.002,12.002,0,0,0-8.48535,3.51465l-24,24a12.00039,12.00039,0,0,0,4.63965,19.85254L64.23413,191.6123l12.28906,36.24122a12,12,0,0,0,19.84961,4.63183l24-24A12.00078,12.00078,0,0,0,123.88745,200V181.31055L133.947,172.1709l40.03223,58.59863a11.99985,11.99985,0,0,0,18.39355,1.71582l24-24a11.99872,11.99872,0,0,0,2.792-12.58594Zm-4.23487,89.95019L145.988,147.251a12,12,0,0,0-17.97753-2.11231l-24.19239,21.97949A12.00061,12.00061,0,0,0,99.88745,176v19.0293l-6.835,6.835L85.05932,178.292a11.99954,11.99954,0,0,0-7.51855-7.51367L54.033,162.8252,60.85815,156h19.0293a12.002,12.002,0,0,0,8.48535-3.51465l24-24a12.00008,12.00008,0,0,0-1.8291-18.46973L50.72339,70.13477l8.24511-8.24512,80.81836,29.38769a11.996,11.996,0,0,0,12.82031-3.0332l31.71045-33.53906c.022-.02344.04444-.04688.06592-.07031A12.0004,12.0004,0,0,1,200.3728,72.48535c-.17676.17676-.34619.35645-.50928.541l-32.18994,30.22558a11.9989,11.9989,0,0,0-3.06348,12.84864L193.99829,196.919Z"></path>
#>   <title>airplane-tilt-bold</title>
#> </svg>
ph("house")
#> <svg xmlns="http://www.w3.org/2000/svg" viewbox="0 0 256 256" class="phosphor-svg" height="1.33em" fill="currentColor" style="vertical-align:-0.25em;">
#>   <path d="M208,221.999h-.00195l-48.00586-.00634a14.01594,14.01594,0,0,1-13.99805-14v-48.001a2.00229,2.00229,0,0,0-2-2h-32a2.00229,2.00229,0,0,0-2,2v48.001a14.01594,14.01594,0,0,1-13.99805,14L48.002,221.999A13.99944,13.99944,0,0,1,34,207.999V115.53857a14.02623,14.02623,0,0,1,4.582-10.35888l79.99414-72.73438a13.94491,13.94491,0,0,1,18.83594-.001L217.418,105.18018A14.03019,14.03019,0,0,1,222,115.53906v92.46a13.99944,13.99944,0,0,1-14,14ZM111.99414,145.9917h32a14.01572,14.01572,0,0,1,14,14v48.001a2.00261,2.00261,0,0,0,2,2L208,209.999h0a2.00051,2.00051,0,0,0,2-2v-92.46a2.00592,2.00592,0,0,0-.65527-1.48047L129.33984,41.32373a1.994,1.994,0,0,0-2.6914.00049L46.6543,114.05908A2.00362,2.00362,0,0,0,46,115.53857V207.999a2.00051,2.00051,0,0,0,2,2l47.99414-.00634a2.00261,2.00261,0,0,0,2-2v-48.001A14.01572,14.01572,0,0,1,111.99414,145.9917Z"></path>
#>   <title>house-light</title>
#> </svg>
ph("house", fill = "#F45B69")
#> <svg xmlns="http://www.w3.org/2000/svg" viewbox="0 0 256 256" class="phosphor-svg" height="1.33em" fill="#F45B69" style="vertical-align:-0.25em;">
#>   <path d="M208,221.999h-.00195l-48.00586-.00634a14.01594,14.01594,0,0,1-13.99805-14v-48.001a2.00229,2.00229,0,0,0-2-2h-32a2.00229,2.00229,0,0,0-2,2v48.001a14.01594,14.01594,0,0,1-13.99805,14L48.002,221.999A13.99944,13.99944,0,0,1,34,207.999V115.53857a14.02623,14.02623,0,0,1,4.582-10.35888l79.99414-72.73438a13.94491,13.94491,0,0,1,18.83594-.001L217.418,105.18018A14.03019,14.03019,0,0,1,222,115.53906v92.46a13.99944,13.99944,0,0,1-14,14ZM111.99414,145.9917h32a14.01572,14.01572,0,0,1,14,14v48.001a2.00261,2.00261,0,0,0,2,2L208,209.999h0a2.00051,2.00051,0,0,0,2-2v-92.46a2.00592,2.00592,0,0,0-.65527-1.48047L129.33984,41.32373a1.994,1.994,0,0,0-2.6914.00049L46.6543,114.05908A2.00362,2.00362,0,0,0,46,115.53857V207.999a2.00051,2.00051,0,0,0,2,2l47.99414-.00634a2.00261,2.00261,0,0,0,2-2v-48.001A14.01572,14.01572,0,0,1,111.99414,145.9917Z"></path>
#>   <title>house-light</title>
#> </svg>
ph("cake", rotate = 45)
#> <svg xmlns="http://www.w3.org/2000/svg" viewbox="0 0 256 256" class="phosphor-svg" height="1.33em" fill="currentColor" style="vertical-align:-0.25em;transform:rotate(45deg);">
#>   <path d="M206,82H134V68.11426c11.5813-4.875,18.34924-11.94922,20.10645-21.0835,3.90722-20.30859-19.49805-41.24707-22.18067-43.56787a5.99953,5.99953,0,0,0-7.7666-.07227,95.564,95.564,0,0,0-12.63379,13.228c-8.63281,11.03858-12.14746,21.20362-10.44824,30.21192C102.82788,56.10693,109.871,63.26758,122,68.1709V82H50a30.03406,30.03406,0,0,0-30,30v13.32812a41.20547,41.20547,0,0,0,12.30957,29.22754c.55066.54053,1.11646,1.05762,1.69043,1.56153V208a14.01572,14.01572,0,0,0,14,14H208a14.01572,14.01572,0,0,0,14-14V156.11621c.574-.50391,1.14-1.02051,1.69043-1.56055A41.20547,41.20547,0,0,0,236,125.32812V112A30.03406,30.03406,0,0,0,206,82ZM112.87109,44.61621c-1.75976-9.27539,8.25293-21.60351,15.041-28.37353a82.227,82.227,0,0,1,6.78711,7.92041c4.20605,5.63476,8.9414,13.78466,7.62109,20.6123-1.01465,5.248-5.84375,9.5669-14.36035,12.85254C118.95117,54.29688,113.87793,49.9248,112.87109,44.61621ZM210,208a2.0026,2.0026,0,0,1-2,2H48a2.0026,2.0026,0,0,1-2-2V163.396a39.129,39.129,0,0,0,14.77832,2.59668,40.112,40.112,0,0,0,33.18933-19.0044,39.966,39.966,0,0,0,68.0647,0,40.112,40.112,0,0,0,33.18933,19.0044c.25488.00488.50781.00732.76172.00732A38.99167,38.99167,0,0,0,210,163.39453Zm14-82.67188a29.13741,29.13741,0,0,1-8.71289,20.66163,27.43979,27.43979,0,0,1-19.83594,8.00488A28.13635,28.13635,0,0,1,168,126a6,6,0,0,0-12,0,28,28,0,0,1-56,0,6,6,0,0,0-12,0,28.13653,28.13653,0,0,1-27.45215,27.99463,27.46611,27.46611,0,0,1-19.835-8.00488A29.13741,29.13741,0,0,1,32,125.32812V112A18.02063,18.02063,0,0,1,50,94H206a18.02063,18.02063,0,0,1,18,18Z"></path>
#>   <title>cake-light</title>
#> </svg>
ph("airplane-tilt", height = 128)
#> <svg xmlns="http://www.w3.org/2000/svg" viewbox="0 0 256 256" class="phosphor-svg" height="128px" fill="currentColor" style="vertical-align:-0.25em;">
#>   <path d="M214.66089,83.53027a30.00065,30.00065,0,0,0-43.98633-40.74414,6.044,6.044,0,0,0-.811,1.00684l-27.6289,29.22168L57.93823,42.36133a5.998,5.998,0,0,0-6.29346,1.39648l-24,24a5.99966,5.99966,0,0,0,.91455,9.23438l65.91016,43.94043L77.4021,138H55.88745a5.99782,5.99782,0,0,0-4.24268,1.75781l-24,24a5.99956,5.99956,0,0,0,2.31983,9.92578l39,13.19532,13.24072,39.04785a5.99955,5.99955,0,0,0,9.9248,2.31543l24-24A5.99653,5.99653,0,0,0,117.88745,200V178.65527l17.126-15.55957,43.91993,64.28907a6.00027,6.00027,0,0,0,9.19677.85742l24-24a5.99849,5.99849,0,0,0,1.396-6.293L182.87768,113.667l30.76856-28.89062A6.00123,6.00123,0,0,0,214.66089,83.53027ZM204.615,76.72852q-.21534.21532-.40528.44726L171.78,107.626a6,6,0,0,0-1.53125,6.4248l30.69433,84.4082L184.7605,214.6416l-43.72657-64.00586a6,6,0,0,0-8.98926-1.05566l-24.19189,21.97949A5.99857,5.99857,0,0,0,105.88745,176v21.51465L90.47,212.93262,79.37719,180.21875a5.99977,5.99977,0,0,0-3.75927-3.75684L42.9602,165.41211,58.3728,150H79.88745a5.99778,5.99778,0,0,0,4.24267-1.75781l24-24a5.99965,5.99965,0,0,0-.91455-9.23438L41.30542,71.06738,57.428,54.94434l84.40918,30.69433a6.0016,6.0016,0,0,0,6.41015-1.5166L179.95776,50.583c.08545-.08985.16748-.18262.24609-.27637A18.00028,18.00028,0,0,1,204.615,76.72852Z"></path>
#>   <title>airplane-tilt-light</title>
#> </svg>
ph("airplane-tilt", height = 128, weight = "fill")
#> <svg xmlns="http://www.w3.org/2000/svg" viewbox="0 0 256 256" class="phosphor-svg" height="128px" fill="currentColor" style="vertical-align:-0.25em;">
#>   <path d="M215.01538,86.23437l-29.80762,27.98829,30.19824,83.043a8.00184,8.00184,0,0,1-1.86181,8.39161l-24,24A7.99924,7.99924,0,0,1,177.282,228.5127l-42.58691-62.33692-18.80762,17.33106V204a8.0008,8.0008,0,0,1-2.34326,5.65723l-20,20a7.99986,7.99986,0,0,1-13.23291-3.08789L67.38745,188.457l-38.064-12.87891a8.00081,8.00081,0,0,1-3.09326-13.23535l20-20A8.00235,8.00235,0,0,1,51.88745,140H72.574l18.75635-18.75684L27.45,78.65625a8.00073,8.00073,0,0,1-1.21973-12.31348l24-24a8.003,8.003,0,0,1,8.39111-1.86132l83.0625,30.20507,26.606-28.14062a7.98306,7.98306,0,0,1,.97071-1.17383,32.0006,32.0006,0,0,1,47.01806,43.34375A8.03536,8.03536,0,0,1,215.01538,86.23437Z"></path>
#>   <title>airplane-tilt-fill</title>
#> </svg>
ph("airplane-tilt", height = 128, weight = "fill", fill = "steelblue")
#> <svg xmlns="http://www.w3.org/2000/svg" viewbox="0 0 256 256" class="phosphor-svg" height="128px" fill="steelblue" style="vertical-align:-0.25em;">
#>   <path d="M215.01538,86.23437l-29.80762,27.98829,30.19824,83.043a8.00184,8.00184,0,0,1-1.86181,8.39161l-24,24A7.99924,7.99924,0,0,1,177.282,228.5127l-42.58691-62.33692-18.80762,17.33106V204a8.0008,8.0008,0,0,1-2.34326,5.65723l-20,20a7.99986,7.99986,0,0,1-13.23291-3.08789L67.38745,188.457l-38.064-12.87891a8.00081,8.00081,0,0,1-3.09326-13.23535l20-20A8.00235,8.00235,0,0,1,51.88745,140H72.574l18.75635-18.75684L27.45,78.65625a8.00073,8.00073,0,0,1-1.21973-12.31348l24-24a8.003,8.003,0,0,1,8.39111-1.86132l83.0625,30.20507,26.606-28.14062a7.98306,7.98306,0,0,1,.97071-1.17383,32.0006,32.0006,0,0,1,47.01806,43.34375A8.03536,8.03536,0,0,1,215.01538,86.23437Z"></path>
#>   <title>airplane-tilt-fill</title>
#> </svg>